+++ core/src/de/cubeisland/games/component/PathFollower.java
@@ -4,7 +4,7 @@ import de.cubeisland.games.entity.Entity;
 
 public class PathFollower extends Component<Entity> {
     @Override
-    public void update(int delta) {
+    public void update(float delta) {
 
     }
 }
